--- pigar/_version.py
@@ -1,6 +1,0 @@
-# -*- coding: utf-8 -*-
-
-from __future__ import print_function, division, absolute_import
-
-version = __version__ = '0.9.2'
-version_info = [int(num) for num in version.split('.')]
