+++ pigar/tests/imports_example/mainfoobar.py
@@ -1,0 +1,5 @@
+from subbar.bar import bar
+import mainfoobar
+
+if __name__ == '__main__':
+    bar()
