+++ pigar/tests/imports_example/subbar/bar.py
@@ -1,0 +1,5 @@
+from subfoo.foo import foo
+
+
+def bar():
+    foo()
