+++ pigar/tests/imports_example/subfoo/foo.py
@@ -1,0 +1,2 @@
+def foo():
+    print("FOO")
